refactor(tropic01_util): separate firmware source code into multiple files
