+++ examples/linux/usb_devkit/tropic01_util/firmware/Inc/app_cmd/pin_set.h
@@ -1,0 +1,8 @@
+#ifndef PIN_SET_H
+#define PIN_SET_H
+
+#include "usb_devkit_messages.pb.h"
+
+void pin_set(const PinSetCmd *cmd, AppResp *resp);
+
+#endif  // PIN_SET_H
