+++ examples/linux/usb_devkit/tropic01_util/firmware/Inc/main.h
@@ -58,7 +58,6 @@ extern RNG_HandleTypeDef hrng;
 extern SPI_HandleTypeDef *hspi1;
 extern CRC_HandleTypeDef hcrc;
 
-extern bool auto_cs_mode;
 extern lt_handle_t lt_handle;
 /* USER CODE END EV */
