+++ examples/linux/usb_devkit/tropic01_util/firmware/Inc/raw_cmd/get_gpo.h
@@ -1,0 +1,8 @@
+#ifndef GET_GPO_H
+#define GET_GPO_H
+
+#include "usb_devkit_messages.pb.h"
+
+void get_gpo(const GetGpoCmd *cmd, GetGpoResp *resp);
+
+#endif  // GET_GPO_H
