+++ examples/linux/usb_devkit/tropic01_util/firmware/Inc/raw_cmd/set_tr01_pwr.h
@@ -1,0 +1,8 @@
+#ifndef SET_TR01_PWR_H
+#define SET_TR01_PWR_H
+
+#include "usb_devkit_messages.pb.h"
+
+void set_tr01_pwr(const SetTr01PwrCmd *cmd, SetTr01PwrResp *resp);
+
+#endif  // SET_TR01_PWR_H
